+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2.0.0.html) and
 - Deprecated methods
 - `SmartEnum` (use `CleverEnum` instead)
 - Memory cache class
+- `FilesysTools.new_webresource`
 
 ### Changed